+++ python_pipeline/requirements.txt
@@ -6,3 +6,4 @@ jsnapy==1.3.8
 lxml
 tabulate
 psutil
+colorama
