+++ python_pipeline/run_jsnapy_tests/cache/172.27.200.200_22_snap_temp_NGzivS58rTvXa5cT3q4LuQ==_get_interface_information.xml
@@ -1,0 +1,205 @@
+<interface-information style="terse"><physical-interface><name>
+ge-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+gr-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ip-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+lsq-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+lt-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+mt-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+sp-0/0/0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/1
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/2
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/3
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/4
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/5
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/6
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+ge-0/0/7
+</name><admin-status>
+up
+</admin-status><oper-status>
+down
+</oper-status></physical-interface><physical-interface><name>
+dl0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+esi
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+fti0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+fxp2
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+gre
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ipip
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+irb
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+jsrv
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+lo0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+lsi
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+mtun
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+pimd
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+pime
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+pp0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ppd0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+ppe0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+rbeb
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+st0
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+tap
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface><physical-interface><name>
+vtep
+</name><admin-status>
+up
+</admin-status><oper-status>
+up
+</oper-status></physical-interface></interface-information>
