+++ python_pipeline/run_jsnapy_tests/cache/172.27.200.200_22_snap_temp_get_software_information.xml
@@ -1,0 +1,1 @@
+<software-information><host-name>MLRENGBSMSRX01</host-name><product-model>srx320-poe</product-model><product-name>srx320-poe</product-name><jsr/><junos-version>24.2R1-S2.5</junos-version><package-information><name>junos</name><comment>JUNOS Software Release [24.2R1-S2.5]</comment></package-information></software-information>
