+++ python_pipeline/run_jsnapy_tests/cache/172.27.200.200_22_snap_temp_yvtTohF6wjJx4gQV0h_L4A==_get_bgp_neighbor_information.xml
@@ -1,0 +1,1 @@
+<bgp-information><bgp-peer style="detail"><peer-address>10.100.250.2+49927</peer-address><peer-as>64790</peer-as><local-address>10.100.250.1+179</local-address><local-as>64790</local-as><peer-group>iBGP-IPV4-UNICAST</peer-group><peer-cfg-rti>master</peer-cfg-rti><peer-fwd-rti>master</peer-fwd-rti><peer-type>Internal</peer-type><peer-state>Established</peer-state><peer-flags>Sync</peer-flags><rsync-flags/><last-state>OpenConfirm</last-state><last-event>RecvKeepAlive</last-event><last-error>None</last-error><bgp-option-information><bgp-options>LocalAddress Refresh</bgp-options><bgp-options2/><bgp-options-extended>GracefulShutdownRcv</bgp-options-extended><bgp-options-extended2/><local-address>10.100.250.1</local-address><holdtime>90</holdtime><preference>170</preference><gshut-recv-local-preference>0</gshut-recv-local-preference></bgp-option-information><flap-count>0</flap-count><recv-ebgp-origin-validation-state>Accept</recv-ebgp-origin-validation-state><bgp-malformed-attributes><malformed-information><log-interval>300</log-interval><route-limit>1000</route-limit></malformed-information></bgp-malformed-attributes><peer-id>10.100.250.2</peer-id><local-id>10.100.250.1</local-id><active-holdtime>90</active-holdtime><keepalive-interval>30</keepalive-interval><group-index>0</group-index><peer-index>0</peer-index><snmp-index>0</snmp-index><bgp-peer-iosession><iosession-thread-name>bgpio-0</iosession-thread-name><iosession-state>Enabled</iosession-state></bgp-peer-iosession><bgp-bfd><bfd-configuration-state>disabled</bfd-configuration-state><bfd-operational-state>down</bfd-operational-state></bgp-bfd><peer-restart-nlri-configured>inet-unicast</peer-restart-nlri-configured><nlri-type-peer>inet-unicast</nlri-type-peer><nlri-type-session>inet-unicast</nlri-type-session><peer-refresh-capability>2</peer-refresh-capability><peer-stale-route-time-configured>300</peer-stale-route-time-configured><peer-no-restart/><peer-restart-nlri-negotiated>inet-unicast</peer-restart-nlri-negotiated><peer-end-of-rib-received>inet-unicast</peer-end-of-rib-received><peer-end-of-rib-sent>inet-unicast</peer-end-of-rib-sent><peer-end-of-rib-scheduled/><peer-no-llgr-helper/><peer-4byte-as-capability-advertised>64790</peer-4byte-as-capability-advertised><peer-addpath-not-supported/><local-ext-nh-color-nlri>inet-unicast</local-ext-nh-color-nlri><bgp-rib style="detail"><name>inet.0</name><rib-bit>20000</rib-bit><bgp-rib-state>BGP restart is complete</bgp-rib-state><send-state>in sync</send-state><active-prefix-count>0</active-prefix-count><received-prefix-count>0</received-prefix-count><accepted-prefix-count>0</accepted-prefix-count><suppressed-prefix-count>0</suppressed-prefix-count><advertised-prefix-count>0</advertised-prefix-count></bgp-rib><last-received>17</last-received><last-sent>24</last-sent><last-checked>157789</last-checked><input-messages>5851</input-messages><input-updates>1</input-updates><input-refreshes>0</input-refreshes><input-octets>111213</input-octets><output-messages>5854</output-messages><output-updates>0</output-updates><output-refreshes>0</output-refreshes><output-octets>111230</output-octets><bgp-output-queue><number>1</number><count>0</count><table-name>inet.0</table-name><rib-adv-nlri>inet-unicast</rib-adv-nlri></bgp-output-queue></bgp-peer></bgp-information>
